Add self reference to rest of firewall rules
As previous commit, we add the self rule to the rest of rules.

The CF components would need to connect to the endpoints of CF, both
private or public (e.g. loggregator checking UAA credentials).

Because that the bosh-lite VM must be able to access itself, which is
achieved by defining self in the security group.